requirements.txt
rich==10.6.0
urllib3==1.26.8
requests==2.27.1
python-telegram-bot==13.11
pandas==1.2.2
python-dotenv==0.19.2
